Extract shared response/error helpers to eliminate route duplication
Both system-logs route handlers had identical response-building,
success logging, and error-handling blocks. Extract to helpers.js
to satisfy SonarCloud duplication check.
