+++ src/routes/v1/system-logs/search.js
@@ -1,18 +1,15 @@
 import Joi from 'joi'
-import Boom from '@hapi/boom'
-import { StatusCodes } from 'http-status-codes'
 
-import {
-  LOGGING_EVENT_ACTIONS,
-  LOGGING_EVENT_CATEGORIES
-} from '#common/enums/index.js'
 import { ROLES } from '#common/helpers/auth/constants.js'
 import { getAuthConfig } from '#common/helpers/auth/get-auth-config.js'
+import {
+  DEFAULT_LIMIT,
+  MAX_LIMIT,
+  handleSystemLogsError,
+  respondWithSystemLogs
+} from './helpers.js'
 
 /** @typedef {import('#repositories/system-logs/port.js').SystemLogsRepository} SystemLogsRepository */
-
-const DEFAULT_LIMIT = 50
-const MAX_LIMIT = 500
 
 const systemLogsPath = '/v1/system-logs'
 
@@ -43,39 +40,9 @@ export const systemLogsGet = {
         cursor
       })
 
-      const response = {
-        systemLogs: result.systemLogs,
-        hasMore: result.hasMore
-      }
-
-      if (result.nextCursor) {
-        response.nextCursor = result.nextCursor
-      }
-
-      logger.info({
-        message: `Listed ${result.systemLogs.length} system logs`,
-        event: {
-          category: LOGGING_EVENT_CATEGORIES.SERVER,
-          action: LOGGING_EVENT_ACTIONS.REQUEST_SUCCESS
-        }
-      })
-
-      return h.response(response).code(StatusCodes.OK)
+      return respondWithSystemLogs(result, h, logger)
     } catch (error) {
-      if (error.isBoom) {
-        throw error
-      }
-
-      logger.error({
-        err: error,
-        message: `Failure on ${systemLogsPath}`,
-        event: {
-          category: LOGGING_EVENT_CATEGORIES.SERVER,
-          action: LOGGING_EVENT_ACTIONS.RESPONSE_FAILURE
-        }
-      })
-
-      throw Boom.badImplementation(`Failure on ${systemLogsPath}`)
+      handleSystemLogsError(error, logger, systemLogsPath)
     }
   }
 }
